[TASK] Move classes into new locations
move classes from Phile namespace into Phile\Core, mark all these classes
as deprecated and add a compatibility class which will be removed in 1.6
and change unit tests

Releases: 1.4
